Animación pista de carreras!
Se agrego el contador de distancia y una animación de pista de carrera cuando se esta en programa manual.
Ademas cuando se esta en el algún programa, la barra de velocidad actual parpadea un par de veces antes de cambiar a la próxima.

+++ Inc/main.h
@@ -191,7 +191,7 @@ typedef struct{
 	uint16_t timmer;
 	angle_e inclinacion;
 	float distancia;
-	float calorias;
+	uint16_t calorias;
 	bool viento;
 }maquina_s;
